Implement algorithm making bounds of convex hexagon tight
Every convex hexagonal shape can be described by up to 6 half planes. It
is, however, possible that not all of them actually contribute to the
shape, since the 2 outer of any 3 adjacent planes might meet in a point
that leaves the middle plane completely outside of the shape of the
polygon.
This algorithm takes all 6 half planes, and makes sure that, in the
described case, the middle plane touches the two adjacent planes in a
single point.
Note that there still are configurations where the 6 half planes do not
designate a convex polygon because no points fulfill all 6 requirements.